Rename background to coverage.
