+++ tests/models/test_customlist.py
@@ -7,7 +7,7 @@ from nose.tools import (
 import datetime
 from .. import DatabaseTest
 from ...model import get_one_or_create
-from ...model.background import WorkCoverageRecord
+from ...model.coverage import WorkCoverageRecord
 from ...model.customlist import (
     CustomList,
     CustomListEntry,
